Fix and refactor MC_HOSTS_ env handling code
Existing MC_HOSTS_<alias> was incorrectly placed
inside s3Config initialization code path which
leads to a situation where we loose the URL path
information.

Fixes #2364
