+++ cmd/utils.go
@@ -21,13 +21,11 @@ import (
 	"fmt"
 	"io"
 	"math/rand"
-	"net/url"
 	"os"
 	"runtime"
 	"strings"
 	"time"
 
-	"github.com/minio/cli"
 	"github.com/minio/mc/pkg/console"
 	"github.com/minio/mc/pkg/probe"
 )
@@ -97,69 +95,9 @@ func splitStr(path, sep string, n int) []string {
 	return splits
 }
 
-// parse url usually obtained from env.
-func parseEnvURL(envURL string) (*url.URL, string, string, *probe.Error) {
-	u, e := url.Parse(envURL)
-	if e != nil {
-		return nil, "", "", probe.NewError(e).Trace(envURL)
-	}
-
-	var accessKey, secretKey string
-	// Check if username:password is provided in URL, with no
-	// access keys or secret we proceed and perform anonymous
-	// requests.
-	if u.User != nil {
-		accessKey = u.User.Username()
-		secretKey, _ = u.User.Password()
-	}
-
-	// Look for if URL has invalid values and return error.
-	if !((u.Scheme == "http" || u.Scheme == "https") &&
-		(u.Path == "/" || u.Path == "") && u.Opaque == "" &&
-		u.ForceQuery == false && u.RawQuery == "" && u.Fragment == "") {
-		return nil, "", "", errInvalidArgument().Trace(u.String())
-	}
-
-	// Now that we have validated the URL to be in expected style.
-	u.User = nil
-
-	return u, accessKey, secretKey, nil
-}
-
-func buildS3ConfigFromEnv(envURL string) (*Config, *probe.Error) {
-	u, accessKey, secretKey, err := parseEnvURL(envURL)
-	if err != nil {
-		return nil, err.Trace(envURL)
-	}
-
-	s3Config, err := newS3Config(cli.Args{u.String(), accessKey, secretKey})
-	if err != nil {
-		return nil, err.Trace(envURL)
-	}
-
-	s3Config.AppName = "mc"
-	s3Config.AppVersion = Version
-	s3Config.AppComments = []string{os.Args[0], runtime.GOOS, runtime.GOARCH}
-	s3Config.Debug = globalDebug
-	s3Config.Insecure = globalInsecure
-
-	return s3Config, nil
-}
-
-const mcEnvHostsPrefix = "MC_HOSTS_"
-
 // buildS3Config fetches config related to the specified alias
 // to create a new config structure
-func buildS3Config(alias, urlStr string) (*Config, *probe.Error) {
-	// Check if we can get the values from ENV, if not use from `config.json`.
-	if alias == "" {
-		alias, _ = url2Alias(urlStr)
-	}
-	if envConfig, ok := os.LookupEnv(mcEnvHostsPrefix + alias); ok {
-		return buildS3ConfigFromEnv(envConfig)
-	}
-
-	hostCfg := mustGetHostConfig(alias)
+func buildS3Config(alias, urlStr string, hostCfg *hostConfigV8) (*Config, *probe.Error) {
 	if hostCfg == nil {
 		return nil, probe.NewError(fmt.Errorf("The specified alias: %s not found", urlStr))
 	}
